// src/redux/popup/types.ts
export interface PopupState {
  activeDate: number;
  isOpenDate: boolean;
  activeCallType: number;
  isOpenCallType: boolean;
  activeWorkers: number;
  isOpenWorkers: boolean;
  activeCalls: number;
  isOpenCalls: boolean;
  activeSources: number;
  isOpenSources: boolean;
  activeEstimates: number;
  isOpenEstimates: boolean;
  activeErrors: number;
  isOpenErrors: boolean;
}